# backend/test_api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Edge Case Test API",
    "version": "1.0.0"
  },
  "paths": {
    "/gadgets": {
      "get": {
        "tags": [
          "gadgets"
        ],
        "summary": "List Gadgets",
        "operationId": "list_gadgets_gadgets_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StatusEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Gadget"
                  },
                  "title": "Response List Gadgets Gadgets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "gadgets"
        ],
        "summary": "Create Gadget",
        "operationId": "create_gadget_gadgets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GadgetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gadgets/{gadget_id}": {
      "get": {
        "tags": [
          "gadgets"
        ],
        "summary": "Get Gadget",
        "operationId": "get_gadget_gadgets__gadget_id__get",
        "parameters": [
          {
            "name": "gadget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gadget Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "gadgets"
        ],
        "summary": "Replace Gadget",
        "operationId": "replace_gadget_gadgets__gadget_id__put",
        "parameters": [
          {
            "name": "gadget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gadget Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GadgetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "gadgets"
        ],
        "summary": "Update Gadget",
        "operationId": "update_gadget_gadgets__gadget_id__patch",
        "parameters": [
          {
            "name": "gadget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gadget Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GadgetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gadgets"
        ],
        "summary": "Delete Gadget",
        "operationId": "delete_gadget_gadgets__gadget_id__delete",
        "parameters": [
          {
            "name": "gadget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gadget Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/containers/{container_id}": {
      "get": {
        "tags": [
          "containers"
        ],
        "summary": "Get Container",
        "operationId": "get_container_containers__container_id__get",
        "parameters": [
          {
            "name": "container_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Container Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GadgetContainer"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gadgets/{gadget_id}/attachments": {
      "post": {
        "tags": [
          "gadgets"
        ],
        "summary": "Upload Attachment",
        "operationId": "upload_attachment_gadgets__gadget_id__attachments_post",
        "parameters": [
          {
            "name": "gadget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Gadget Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_attachment_gadgets__gadget_id__attachments_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_attachment_gadgets__gadget_id__attachments_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "file",
          "label"
        ],
        "title": "Body_upload_attachment_gadgets__gadget_id__attachments_post"
      },
      "ErrorCodeEnum": {
        "type": "string",
        "enum": [
          "UNKNOWN",
          "NOT_FOUND",
          "INVALID"
        ],
        "title": "ErrorCodeEnum"
      },
      "ErrorResponse": {
        "properties": {
          "error_code": {
            "$ref": "#/components/schemas/ErrorCodeEnum",
            "default": "UNKNOWN"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "Gadget": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/StatusEnum"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum",
            "default": "low"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "ratio": {
            "type": "number",
            "title": "Ratio"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Metadata",
            "default": {}
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra",
            "default": {}
          },
          "html_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Url"
          },
          "default": {
            "type": "integer",
            "title": "Default",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "name",
          "status",
          "ratio",
          "created_at"
        ],
        "title": "Gadget"
      },
      "GadgetContainer": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "primary": {
            "$ref": "#/components/schemas/Gadget"
          },
          "gadgets": {
            "items": {
              "$ref": "#/components/schemas/Gadget"
            },
            "type": "array",
            "title": "Gadgets"
          },
          "gadget_map": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Gadget"
            },
            "type": "object",
            "title": "Gadget Map"
          }
        },
        "type": "object",
        "required": [
          "id",
          "primary",
          "gadgets",
          "gadget_map"
        ],
        "title": "GadgetContainer"
      },
      "GadgetCreate": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/StatusEnum"
          },
          "priority": {
            "$ref": "#/components/schemas/PriorityEnum",
            "default": "low"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "name",
          "status"
        ],
        "title": "GadgetCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PriorityEnum": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "PriorityEnum"
      },
      "StatusEnum": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "archived"
        ],
        "title": "StatusEnum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}